Refactor: improved package structure with cli, core, utils and integrations folders
