+++ .pre-commit-config.yaml
@@ -3,7 +3,7 @@ repos:
     hooks:
     -   id: update-command-docs
         name: Update command docs
-        entry: poetry run python dbtwiz/helpers/typer_docs.py
+        entry: poetry run python dbtwiz/utils/typer_docs.py
         language: system
         pass_filenames: false
         always_run: false
