+++ dbtwiz/admin/backfill.py
@@ -8,8 +8,8 @@ from textwrap import dedent
 from jinja2 import Template
 
 from dbtwiz.config.project import project_config, project_dbtwiz_path
-from dbtwiz.gcp.auth import ensure_auth
-from dbtwiz.helpers.logger import debug, info
+from dbtwiz.integrations.gcp_auth import ensure_auth
+from dbtwiz.utils.logger import debug, info
 
 MAX_CONCURRENT_TASKS = 8
 YAML_FILE = project_dbtwiz_path("backfill-cloudrun.yaml")
