+++ dbtwiz/admin/partition.py
@@ -1,8 +1,8 @@
 from typing import Dict, List
 
 from dbtwiz.dbt.manifest import Manifest
-from dbtwiz.gcp.bigquery import BigQueryClient
-from dbtwiz.helpers.logger import info
+from dbtwiz.integrations.bigquery import BigQueryClient
+from dbtwiz.utils.logger import info
 from dbtwiz.ui.interact import multiselect_from_list
