+++ dbtwiz/cli/__init__.py
@@ -1,0 +1,5 @@
+"""CLI-specific functionality for dbtwiz."""
+
+from dbtwiz.cli.main import main
+
+__all__ = ["main"]
