+++ dbtwiz/commands/build.py
@@ -5,8 +5,8 @@ from datetime import date, timedelta
 from dbtwiz.config.project import project_config, project_dbtwiz_path
 from dbtwiz.dbt.manifest import Manifest
 from dbtwiz.dbt.run import invoke
-from dbtwiz.gcp.auth import ensure_auth
-from dbtwiz.helpers.logger import debug, error, info
+from dbtwiz.integrations.gcp_auth import ensure_auth
+from dbtwiz.utils.logger import debug, error, info
 
 LAST_SELECT_FILE = project_dbtwiz_path("last_select.json")
