+++ dbtwiz/commands/test.py
@@ -3,8 +3,8 @@ from datetime import date
 from dbtwiz.config.project import project_config
 from dbtwiz.dbt.manifest import Manifest
 from dbtwiz.dbt.run import invoke
-from dbtwiz.gcp.auth import ensure_auth
-from dbtwiz.helpers.logger import debug, error, info
+from dbtwiz.integrations.gcp_auth import ensure_auth
+from dbtwiz.utils.logger import debug, error, info
 
 
 def test(
