+++ dbtwiz/core/model.py
