+++ dbtwiz/core/project.py
@@ -4,7 +4,7 @@ from typing import Dict, List, Tuple, Union
 
 from dbtwiz.config.project import project_path
 
-from .model import ModelBasePath
+from dbtwiz.core.model import ModelBasePath
 
 
 class Group:
