+++ dbtwiz/utils/__init__.py
@@ -1,0 +1,17 @@
+"""Utility functions for dbtwiz."""
+
+from dbtwiz.utils.logger import debug, info, warn, error, fatal, notice, status
+from dbtwiz.utils.editor import open_in_editor
+from dbtwiz.dbt.support import models_with_local_changes
+
+__all__ = [
+    "debug",
+    "info", 
+    "warn",
+    "error",
+    "fatal",
+    "notice",
+    "status",
+    "open_in_editor",
+    "models_with_local_changes"
+]
